feat: add custom logic for validating a user handle during registration in strapi-plugin-users-permissions/controllers/Auth.js
There are a couple issues with local auth in strapi that I implemented
workarounds for.

First issue is that the sqlite database
apparently doesn't show which unique key in a database entry failed when
there is a duplicate key error. I imagine the ORM includes the field for
other databases and this bug is specific to sqlite. The problem is that
the value always comes back as email since the check fails to include
the word username. The check that is implemented sucks anyway, and
should be more sophisticated, but this will move me forward for my
purposes.

Second issue is that there apparently aren't checks being done on custom
fields I add to the user model (e.g. handle) based on the options I set
in the admin panel. I set a regex and min/max length. I'm not sure if
skipping these settings is Auth module specific or not, but the
workaround in this commit should be enough for me.